Add <TopicsFilter> to filter the table by tag
Modeling after <TextFilter>, this filter class for the table allows
users to hone in on a tag they may want to look at. Not the
be-all-end-all, but this is a start to get the concept of topic tags in
the product.

I took the opportunity to learn a bit about React testing and tried to
write some tests. I feel like this stuff is pretty jank, suggestions
would be appreciated.