Replace memcached with redis
Memcached has a 1 MB object size limit, which we just hit with the
addition of `last_pushed_within` field. To accommodate larger objects,
let's switch to redis.

This commit brings in the `apicache` npm module to handle the caching
via express middleware for us. I removed the `updateCacheTask` as we are
not using the tag functionality and the cache updates are handled
automatically by the middleware.