Only show `Load next 50 projects` button when applicable
The pagination button erroneously shows up when filtering the set of
projects to a small set (it shows despite there not being any more
projects to load).

By comparing the `rows` length (the array of rows after the filters have
been applied) to the pageSize instead, we can get the accurate
understanding of if there are more rows.